+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Changed
   * AZ-908: Use the new data source for CAF naming (instead of resource)
 
+Breaking
+  * AZ-908: Implement CAF naming for Virtual Hub module
+
 # v7.1.0 - 2022-10-28
 
 Changed